+++ Re/Controller.h
@@ -36,7 +36,7 @@ private:
 	//コントローラー振動強さ
 	float shakePower = 0.0f;
 	//コントローラー振動長さ(フレーム数)
-	int shakeTimer = 0;
+	size_t shakeTimer = 0;
 
 	//デッドゾーンに入っているか(DeadRate : デッドゾーン判定の度合い、1.0fだとデフォルト)
 	bool StickInDeadZone(Vector2D thumb, const Vector2D deadRate);
@@ -118,7 +118,7 @@ public:
 	/// </summary>
 	/// <param name="power">振動の強さ0.0f～1.0f</param>
 	/// <param name="span">振動の時間フレーム</param>
-	void ShakeController(const float& power, const int& span);
+	void ShakeController(const float& power, const size_t& span);
 
 	/// <summary>
 	/// Lスティックの傾きの取得
