+++ Re/Matrix.cpp
@@ -38,10 +38,10 @@ Matrix& Matrix::operator*=(const Matrix& m2)
 bool InverseMatrix(const Matrix& mat, Matrix& invMat)
 {
     float sweep[4][8] = { 0 };
-    int k = 0;
+    size_t k = 0;
 
-    for (int y = 0; y < 4; y++) {
-        for (int x = 0; x < 4; x++) {
+    for (size_t y = 0; y < 4; y++) {
+        for (size_t x = 0; x < 4; x++) {
             // sweepの左側に逆行列を求める行列をセット
             sweep[y][x] = mat.m[y][x];
 
@@ -61,10 +61,10 @@ bool InverseMatrix(const Matrix& mat, Matrix& invMat)
 
         /* 最大の絶対値を注目対角成分の絶対値と仮定 */
         double max = fabs(sweep[k][k]);
-        int max_i = k;
+        size_t max_i = k;
 
         /* k列目が最大の絶対値となる行を探す */
-        for (int y = k + 1; y < 4; y++) {
+        for (size_t y = k + 1; y < 4; y++) {
             if (fabs(sweep[y][k]) > max) {
                 max = fabs(sweep[y][k]);
                 max_i = y;
@@ -78,7 +78,7 @@ bool InverseMatrix(const Matrix& mat, Matrix& invMat)
 
         /* 操作（１）：k行目とmax_i行目を入れ替える */
         if (k != max_i) {
-            for (int x = 0; x < 8; x++) {
+            for (size_t x = 0; x < 8; x++) {
                 float tmp = sweep[max_i][x];
                 sweep[max_i][x] = sweep[k][x];
                 sweep[k][x] = tmp;
@@ -89,13 +89,13 @@ bool InverseMatrix(const Matrix& mat, Matrix& invMat)
         float a = 1 / sweep[k][k];
 
         /* 操作（２）：k行目をa倍する */
-        for (int x = 0; x < 8; x++) {
+        for (size_t x = 0; x < 8; x++) {
             /* これによりsweep[k][k]が1になる */
             sweep[k][x] *= a;
         }
 
         /* 操作（３）によりk行目以外の行のk列目を0にする */
-        for (int y = 0; y < 4; y++) {
+        for (size_t y = 0; y < 4; y++) {
             if (y == k) {
                 /* k行目はそのまま */
                 continue;
@@ -104,7 +104,7 @@ bool InverseMatrix(const Matrix& mat, Matrix& invMat)
             /* k行目に掛ける値を求める */
             a = -sweep[y][k];
 
-            for (int x = 0; x < 8; x++) {
+            for (size_t x = 0; x < 8; x++) {
                 /* i行目にk行目をa倍した行を足す */
                 /* これによりsweep[i][k]が0になる */
                 sweep[y][x] += sweep[k][x] * a;
@@ -113,8 +113,8 @@ bool InverseMatrix(const Matrix& mat, Matrix& invMat)
     }
 
     /* sweepの右半分がmatの逆行列 */
-    for (int y = 0; y < 4; y++) {
-        for (int x = 0; x < 4; x++) {
+    for (size_t y = 0; y < 4; y++) {
+        for (size_t x = 0; x < 4; x++) {
             invMat.m[y][x] = sweep[y][4 + x];
         }
     }
