+++ Re/ScissorRect.h
@@ -7,7 +7,7 @@ private:
 	D3D12_RECT scissorRect{};
 public:
 	ScissorRect();
-	void Init(const int left, const int width, const int top, const int height);
+	void Init(const size_t left, const size_t width, const size_t top, const size_t height);
 	void Update(ID3D12GraphicsCommandList* cmdList);
 };
