+++ Re/Window.h
@@ -5,8 +5,8 @@ class Window
 {
 public:
 	// ウィンドウサイズ
-	static const int window_width = 1280; // 横幅
-	static const int window_height = 720; // 縦幅
+	static const size_t window_width = 1280; // 横幅
+	static const size_t window_height = 720; // 縦幅
 private:
 	// ウィンドウクラスの設定
 	WNDCLASSEX w{};
